Completed get function, working on delete

+++ WebAPISample/Controllers/MovieController.cs
@@ -13,23 +13,28 @@ namespace WebAPISample.Controllers
         private ApplicationDbContext db = new ApplicationDbContext();
 
         // GET api/values
-        public IEnumerable<string> Get() // Retrieve all movies from db logic
+        public IHttpActionResult Get() // Retrieve all movies from db logic
         {
-
-            return new string[] { "movie1 string", "movie2 string" };
+            List<Movie> movies = new List<Movie>();
+            foreach(var movie in db.Movies)
+            {
+                movies.Add(movie);
+            }
+            
+            return Ok(movies);
         }
 
         // GET api/values/5
-        public string Get(int id) // Retrieve movie by id from db logic
-        { 
-            Movie movie = db.Movies.Find(id);
-           
-            return "value";
+        public IHttpActionResult Get(int id) // Retrieve movie by id from db logic
+        {
+            var movie = db.Movies.Find(id);
+
+            return Ok(movie);
         }
 
         // POST api/values
         public void Post([FromBody]Movie movie) // Create movie in db logic
-        {  
+        {
             if (ModelState.IsValid)
             {
                 db.Movies.Add(movie);
@@ -47,11 +52,19 @@ namespace WebAPISample.Controllers
         }
 
         // DELETE api/values/5
-        public void Delete(int id) // Delete movie from db logic
+        public IHttpActionResult Delete(int id) // Delete movie from db logic
         {
-            Movie movie = db.Movies.Find(id);
-            db.Movies.Remove(movie);
-            db.SaveChanges();           
+            if (id == 0)
+            {
+                return BadRequest("Not a valid id");
+            }
+            else
+            {
+                var movie = db.Movies.Where(m => m.MovieId == id).FirstOrDefault();
+                db.Movies.Remove(movie);
+                db.SaveChanges();
+                return Ok();
+            }
         }
     }
